Enable HW1 component and add project settings for deployment
Uncommented the HW1 component in App.tsx and added UserType to HW1 array for proper rendering. Further changes were added to the package.json for project deployment configurations and updates were made to the yarn.lock file with dependencies related to deployment. The commit allows the deployment of the application to GitHub Pages.“